code runs but doesnt produce to kafka
