--- K_Way_Merge_Kafka/.gitignore
@@ -1,5 +1,0 @@
-# Ignore IntelliJ IDEA files
-.idea/
-
-# Ignore Visual Studio Code files
-.vscode/
