--- K_Way_Merge_Kafka/KWayMerge/src/main/java/com/example/kafka/kafkaMessage.java
@@ -1,11 +1,0 @@
-package com.example.kafka;
-
-public class kafkaMessage {
-    int seqNum, value;
-    long arrivalTime;
-    public kafkaMessage(int seqNum, int value, long time) {
-        this.seqNum = seqNum;
-        this.value = value;
-        this.arrivalTime = time;
-    }
-}
