--- K_Way_Merge_Kafka/KWayMerge/src/test/java/com/example/kafka/KMergeTest.java
@@ -1,48 +1,0 @@
-package com.example.kafka;
-
-import junit.framework.Test;
-import junit.framework.TestCase;
-import junit.framework.TestSuite;
-import org.apache.kafka.clients.admin.AdminClient;
-import org.apache.kafka.clients.admin.AdminClientConfig;
-import org.apache.kafka.clients.admin.KafkaAdminClient;
-import org.testcontainers.containers.KafkaContainer;
-import org.testcontainers.utility.DockerImageName;
-import org.apache.kafka.clients.admin.NewTopic;
-
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.Properties;
-
-/**
- * Unit test for simple App.
- */
-public class KMergeTest
-    extends TestCase
-{
-    /**
-     * Create the test case
-     *
-     * @param testName name of the test case
-     */
-    public KMergeTest(String testName )
-    {
-        super( testName );
-    }
-
-
-    /**
-     * @return the suite of tests being tested
-     */
-    public static Test suite()
-    {
-        return new TestSuite( KMergeTest.class );
-    }
-
-    /**
-     * Rigourous Test :-)
-     */
-}
-
